// frontend/src/api/models/ZzzOdDirectBackupIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 直控前置保护：进入直控前确保一条龙原生配置已有备份
 */
export type ZzzOdDirectBackupIn = {
    /**
     * 所属脚本ID
     */
    scriptId: string;
};