// frontend/src/api/models/ZzzOdInstanceAddIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 直控：新建一条龙实例（分配最小空闲槽并注册）
 */
export type ZzzOdInstanceAddIn = {
    /**
     * 所属脚本ID
     */
    scriptId: string;
    /**
     * 实例名称（必填，创建后可在重命名中修改）
     */
    name: string;
};