// frontend/src/api/models/ZzzOdInstanceDeleteIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 直控：删除实例（注册表条目 + 实例目录，受 MAS 绑定槽保护）
 */
export type ZzzOdInstanceDeleteIn = {
    /**
     * 所属脚本ID
     */
    scriptId: string;
    /**
     * 目标实例下标
     */
    instanceIdx: number;
};