// frontend/src/api/models/ZzzOdInstanceRenameIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 直控：重命名实例（只改注册表 name，实例目录不变）
 */
export type ZzzOdInstanceRenameIn = {
    /**
     * 所属脚本ID
     */
    scriptId: string;
    /**
     * 目标实例下标
     */
    instanceIdx: number;
    /**
     * 新实例名称
     */
    name: string;
};